+++ x/leverage/keeper/interest.go
@@ -10,7 +10,7 @@ import (
 )
 
 // DeriveBorrowAPY derives the current borrow interest rate on a token denom
-// using its borrow utilization and token-specific params. Returns zero on
+// using its supply utilization and token-specific params. Returns zero on
 // invalid asset.
 func (k Keeper) DeriveBorrowAPY(ctx sdk.Context, denom string) sdk.Dec {
 	token, err := k.GetTokenSettings(ctx, denom)
@@ -23,7 +23,7 @@ func (k Keeper) DeriveBorrowAPY(ctx sdk.Context, denom string) sdk.Dec {
 		return sdk.ZeroDec()
 	}
 
-	utilization := k.ComputeBorrowUtilization(ctx, denom)
+	utilization := k.SupplyUtilization(ctx, denom)
 
 	if utilization.GTE(token.KinkUtilization) {
 		return Interpolate(
@@ -46,7 +46,7 @@ func (k Keeper) DeriveBorrowAPY(ctx sdk.Context, denom string) sdk.Dec {
 }
 
 // DeriveLendAPY derives the current lend interest rate on a token denom
-// using its borrow utilization borrow APY. Returns zero on invalid asset.
+// using its supply utilization borrow APY. Returns zero on invalid asset.
 func (k Keeper) DeriveLendAPY(ctx sdk.Context, denom string) sdk.Dec {
 	token, err := k.GetTokenSettings(ctx, denom)
 	if err != nil {
@@ -54,11 +54,11 @@ func (k Keeper) DeriveLendAPY(ctx sdk.Context, denom string) sdk.Dec {
 	}
 
 	borrowRate := k.DeriveBorrowAPY(ctx, denom)
-	borrowUtilization := k.ComputeBorrowUtilization(ctx, denom)
+	utilization := k.SupplyUtilization(ctx, denom)
 	reduction := k.GetParams(ctx).OracleRewardFactor.Add(token.ReserveFactor)
 
 	// lend APY = borrow APY * utilization, reduced by reserve factor and oracle reward factor
-	return borrowRate.Mul(borrowUtilization).Mul(sdk.OneDec().Sub(reduction))
+	return borrowRate.Mul(utilization).Mul(sdk.OneDec().Sub(reduction))
 }
 
 // AccrueAllInterest is called by EndBlock to update borrow positions.